[tra-16345]Mettre la modale de signature de l'émetteur d'un BSDA de réexpédition / groupement au DSFR
