+++ front/src/dashboard/detail/bsda/BsdaDetailContent.tsx
@@ -30,7 +30,7 @@ import React, { useMemo } from "react";
 import QRCodeIcon from "react-qr-code";
 import { generatePath, useNavigate, useParams } from "react-router-dom";
 import { Tab, TabList, TabPanel, Tabs } from "react-tabs";
-import { InitialBsdas } from "./InitialBsdas";
+import { InitialBsdas } from "../../../Apps/Dashboard/Validation/Bsda/InitialBsdas";
 import { getOperationModeLabel } from "../../../Apps/common/operationModes";
 import EstimatedQuantityTooltip from "../../../common/components/EstimatedQuantityTooltip";
 import { BSDA_VERBOSE_STATUSES } from "@td/constants";
